+++ earthfile2llb/cloneable.go
@@ -1,0 +1,5 @@
+package earthfile2llb
+
+type cloneable[T any] interface {
+	Clone() T
+}