+++ earthfile2llb/interpreter.go
@@ -142,18 +142,11 @@ func (i *Interpreter) handleBlockParallel(ctx context.Context, b spec.Block, sta
 		stmt := b[index]
 		if stmt.Command != nil {
 			switch stmt.Command.Name {
-			case command.Arg, command.Locally, command.From, command.FromDockerfile:
+			case command.Arg, command.Locally, command.From, command.FromDockerfile, command.Let, command.Set:
 				// Cannot do any further parallel builds - these commands need to be
 				// executed to ensure that they don't impact the outcome. As such,
 				// commands following these cannot be executed preemptively.
 				return nil
-			case command.Let, command.Set:
-				if i.converter.ftrs.LetSetBlockParallel {
-					// treat LET/SET the same as ARG if the feature flag is on,
-					// otherwise fallthrough to handle the build
-					return nil
-				}
-				fallthrough
 			case command.Build:
 				err := i.handleBuild(ctx, *stmt.Command, true)
 				if err != nil {
@@ -907,6 +900,7 @@ func (i *Interpreter) handleCopy(ctx context.Context, cmd spec.Command) error {
 		return i.errorf(cmd.SourceLocation, "COPY --from not implemented. Use COPY artifacts form instead")
 	}
 	srcs := args[:len(args)-1]
+	srcArtifacts := make([]domain.Artifact, len(srcs))
 	srcFlagArgs := make([][]string, len(srcs))
 	dest, err := i.expandArgs(ctx, args[len(args)-1], false, false)
 	if err != nil {
@@ -972,6 +966,7 @@ func (i *Interpreter) handleCopy(ctx context.Context, cmd spec.Command) error {
 		// If it parses as an artifact, treat as artifact.
 		if parseErr == nil {
 			srcs[index] = artifactSrc.String()
+			srcArtifacts[index] = artifactSrc
 			allClassical = false
 		} else {
 			expandedSrc, err := i.expandArgs(ctx, src, false, false)
@@ -1018,16 +1013,31 @@ func (i *Interpreter) handleCopy(ctx context.Context, cmd spec.Command) error {
 			if !i.converter.ftrs.PassArgs && opts.PassArgs {
 				return i.errorf(cmd.SourceLocation, "the COPY --pass-args flag must be enabled with the VERSION --pass-args feature flag.")
 			}
-
-			if i.local {
-				err = i.converter.CopyArtifactLocal(ctx, src, dest, platform, allowPrivileged, opts.PassArgs, srcBuildArgs, opts.IsDirCopy)
-				if err != nil {
-					return i.wrapError(err, cmd.SourceLocation, "copy artifact locally")
+			expandedSrcs := []string{src}
+			if strings.Contains(srcArtifacts[index].Target.String(), "*") {
+				if !i.converter.ftrs.WildcardCopy {
+					return i.errorf(cmd.SourceLocation, "wildcard COPY commands are not enabled")
 				}
-			} else {
-				err = i.converter.CopyArtifact(ctx, src, dest, platform, allowPrivileged, opts.PassArgs, srcBuildArgs, opts.IsDirCopy, opts.KeepTs, opts.KeepOwn, expandedChown, fileModeParsed, opts.IfExists, opts.SymlinkNoFollow)
+				expandedArtifacts, err := i.converter.ExpandWildcardArtifacts(ctx, srcArtifacts[index])
 				if err != nil {
-					return i.wrapError(err, cmd.SourceLocation, "copy artifact")
+					return i.wrapError(err, cmd.SourceLocation, "failed to expand wildcard COPY %q", srcArtifacts[index].Target.String())
+				}
+				expandedSrcs = make([]string, 0, len(expandedArtifacts))
+				for _, artifact := range expandedArtifacts {
+					expandedSrcs = append(expandedSrcs, artifact.String())
+				}
+			}
+			for _, expandedSrc := range expandedSrcs {
+				if i.local {
+					err = i.converter.CopyArtifactLocal(ctx, expandedSrc, dest, platform, allowPrivileged, opts.PassArgs, srcBuildArgs, opts.IsDirCopy)
+					if err != nil {
+						return i.wrapError(err, cmd.SourceLocation, "copy artifact locally")
+					}
+				} else {
+					err = i.converter.CopyArtifact(ctx, expandedSrc, dest, platform, allowPrivileged, opts.PassArgs, srcBuildArgs, opts.IsDirCopy, opts.KeepTs, opts.KeepOwn, expandedChown, fileModeParsed, opts.IfExists, opts.SymlinkNoFollow)
+					if err != nil {
+						return i.wrapError(err, cmd.SourceLocation, "copy artifact")
+					}
 				}
 			}
 		}
@@ -1284,7 +1294,7 @@ func (i *Interpreter) handleBuild(ctx context.Context, cmd spec.Command, async b
 
 func (i *Interpreter) handleWildcardBuilds(ctx context.Context, fullTargetName string, cmd spec.Command, async bool) error {
 
-	children, err := i.converter.ExpandWildcard(ctx, fullTargetName, cmd)
+	children, err := i.converter.ExpandWildcardCmds(ctx, fullTargetName, cmd)
 	if err != nil {
 		return i.wrapError(err, cmd.SourceLocation, "failed to expand wildcard BUILD %q", fullTargetName)
 	}
@@ -2156,9 +2166,7 @@ func (i *Interpreter) isArgLike(cmd *spec.Command) bool {
 	}
 
 	switch cmd.Name {
-	case command.Let, command.Set:
-		return i.converter.ftrs.LetSetBlockParallel
-	case command.Arg:
+	case command.Arg, command.Let, command.Set:
 		return true
 	}